Move schemas/crud modules out of db package
As discussed in https://github.com/andreagrandi/covid-api/issues/5

This seperates out some stuff that is not really related to the db,
and means that the api layer does not need to interact with the db
directly.

With this architecture, the dependency graph should look something like this

   api -------------
    |               |
    v               v
  services -----> schemas
    |
    v
    db

I think `schemas` and `services` could be modules rather than packages, but
I've made them packages because I'm assuming they'll grow, and to avoid
cluttering the top level package (at the moment this only contains executable
scripts).