Don't import models directly from views
Currently the views module has a side effect of creating the schema,
if it does not already exist. If we move this line elsewhere, then
this we don't need to import the models directly (all interaction
with them is via the crud functions).

Logically I think these belong with the models. This means that the
schema would also be created when running the import script, or
anything else that uses these models.

If we want to make it more explicit we could also move this call to the top
level scripts (app.py and import_data.py) and duplicate the call.