Add dummy tests for requests that read from the db
This is an experiment to see how we can use the "override_dependency"
feature of FastAPI, described here: https://fastapi.tiangolo.com/advanced/testing-dependencies/

I'm importing the TestClient directly from starlette because the fastapi import wasn't working
for some reason. It should be the same thing as fastapi.testclient.

How this works:

1. The test schema is created if it doesn't exist. This mirrors how we create the schema in the app.

2. The test db session is created as a pytest fixture (https://docs.pytest.org/en/latest/fixture.html)

3. The test db session overrides the fastapi dependencies

4. The test db session is automatically rolled back at the end of each test, to allow
   models to be created for the setup of each test.