Get the data from github, instead of a downloaded copy
This uses dataflows to download the data from github, while
caching a copy on disk in case the process needs to be rerun.

I will probably remove this part later, because saving to disk won't work on heroku.

We could rewrite this just using requests, but dataflows seems pretty
convenient. I used it because there is already code that uses it to
clean up this dataset, and I think we could reuse some of this later:
https://github.com/datasets/covid-19/blob/master/process.py