Make the import script idempotent
Before inserting a daily report, check if the daily report already exists.
If it exists, but one of the values is different, update them to match the CSV.

As of today, there are 3430 regions per daily report, so we're executing 6860 queries
per day we want to import. This takes about 35 seconds to run.