Add new fields and a repr method to the sqlalchemy model
The new fields have been added in more recent reports to allow
more granular reporting of US data.

the __repr__ helps debugging when something goes wrong with the import.

I based this __repr__ function on the one in sqlalchemy_utils,
but I didn't think it was worth pulling in another dependency
just for this. This method uses sqlalchemy's inspect function
because reading attributes normally may have the side effect
of loading it on demand.