Add setup.py with entry points for import scripts
Until now, we've run the import script by running it directly. This was
working fine as long as we only used absolute imports, and import things that
only use absolute imports.

Using relative imports (with a leading dot) throws lots of "attempted relative
import beyond top-level package" errors, because the package itself hasn't
actually been imported, so the top level package is actually the script we're
running *within* the package.

I also get import errors when I try to write tests that import the import
script, (i.e. it can't import `db` in this context, since pytest knows it's
part of `covidapi`).

We could move the import script outside the package itself, but there will be
more of these scripts and I don't want to clutter up the root of the project.

Instead, I've created a setup.py so we can make entry points available within
the package that you can run on the command line. When these commands are run,
imports behave more intuitively, and still work when imported from a test.

If you run `python setup.py develop` within a virtualenv, the package sources
will be symlinked instead of copied so the commands will reflect the actual
source we're working on.

On heroku, we can run `python setup.py install` to install the package.