Try to use JH ID to find lookup existing reports
If the import is rerun, we try and update existing reports rather than adding new ones.

To do this, first try to do this by JH_ID. If present this means we've already mapped
this to the locations in the lookup table.

Otherwise, we may not have done this, so fall back to previous method.